+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_03_01_000000) do
+ActiveRecord::Schema[8.1].define(version: 2026_03_02_000000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -121,6 +121,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_03_01_000000) do
     t.text "description"
     t.date "due_date"
     t.string "eligible_reward"
+    t.integer "estimated_daily_task_time"
     t.datetime "hold_until"
     t.bigint "idea_id", null: false
     t.integer "priority"